ADVANCED Project: Final Task: Smart Baggage Tracker using IoT:
Project Summary: In response to the common issue of lost luggage, smart baggage trackers offer a solution by detecting the bag's location in case of theft and sending immediate coordinates to the user's phone. This innovation has significantly reduced the hassles of bag loss in the tourism sector, providing travelers with peace of mind regarding the security of their belongings.